+++ clients/admin-ui/src/features/system/history/modal/SystemHistoryModal.tsx
@@ -11,7 +11,7 @@ import {
   Spacer,
 } from "@fidesui/react";
 
-import { SystemHistory } from "~/types/api/models/SystemHistory";
+import { SystemHistoryResponse } from "~/types/api";
 
 import SelectedHistoryProvider from "./SelectedHistoryContext";
 import SystemDataForm from "./SystemDataForm";
@@ -42,7 +42,7 @@ const getBadges = (before: Record<string, any>, after: Record<string, any>) => {
 };
 
 interface Props {
-  selectedHistory: SystemHistory;
+  selectedHistory: SystemHistoryResponse;
   isOpen: boolean;
   onClose: () => void;
 }